requirements.txt
cffi==1.14.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-pcapng==1.0
scapy==2.4.3
six==1.14.0
